Fire close event before closing and on error
Firing the close event before closing allows for the user to grab the connection's unique ID. This is useful if the user is keeping track of connection based states across the lifecycle of a connection (e.g. session data).

Fire the event on a protocol error as well before closing.