Protocol: Read the whole frame when length is zero.
When the length of the message is zero, the `readFrame` method returns
earlier. However, if a masking key is present in the frame, it is not
read and then still present in the buffer. So the next `readFrame` calls
will be invalidated. To avoid that, the masking key is read, so the
whole frame, each time the `readFrame` method is called.