Connection: If a normal close fails, no error.
When a normal close is attempted and it fails because of several reasons
(it is likely to be because the client might already be disconnected and
then it is not possible to send messages to it), then this situation is
not considered as an error.

Indeed, the client has sent a normal close but it does not listen for
the confirmation from the server. This is a protocol error but happening
on closing.

So instead of getting an exception when closing, catching this exception
later and then firing an `error` event, this makes more sense to catch
the exception, do nothing with it, and fire the expected `close` event.