fix(php) Return type incompatibility.
In the Bin/Client.php class, the return type for the `main` method is `:int`. But at line 79, there is a call to `$this->usage()` which return `void`.

Since the value returned by that method is not compatible with the `main` signature, we get an error like :

```
Uncaught TypeError: Return value of Hoa\Websocket\Bin\Client::main() must be of the type integer, null returned
```